pyproject.toml
[tool.poetry]
name = "habitflow-api"
version = "0.1.0"
description = "API для трекера привычек"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.2"
djangorestframework = "^3.14"
psycopg2-binary = "^2.9"
django-cors-headers = "^4.2"
djangorestframework-simplejwt = "^5.3"
celery = "^5.3"
redis = "^5.0"
python-telegram-bot = "^20.3"
drf-yasg = "^1.21"
python-dotenv = "^1.0"
gunicorn = "^21.2"
whitenoise = "^6.0"
build = "==1.3.0"
cachecontrol = "==0.14.4"
charset-normalizer = "==3.4.4"
cleo = "==2.1.0"
crashtest = "==0.4.1"
cron-descriptor = "==2.0.6"
distlib = "==0.4.0"
django-celery-beat = "==2.8.1"
django-filter = "==25.2"
django-timezone-field = "==7.2.1"
django-celery-results = "==2.6.0"
djangorestframework-stubs = "==3.16.6"
dulwich = "==0.24.10"
execnet = "==2.1.2"
fastjsonschema = "==2.21.2"
filelock = "==3.20.0"
findpython = "==0.7.1"
installer = "==0.7.0"
jaraco-classes = "==3.4.0"
jaraco-context = "==6.0.1"
jaraco-functools = "==4.3.0"
keyring = "==25.7.0"
more-itertools = "==10.8.0"
msgpack = "==1.1.2"
pbs-installer = "==2025.12.2"
pillow = "==12.0.0"
pkginfo = "==1.12.1.2"
drf-spectacular = "==0.27.1"
setuptools = "==70.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
coverage = "^7.4"
black = "^24.0"
flake8 = "^7.0"
mypy = "^1.8"
django-stubs = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 119
skip = ["migrations", "venv", ".venv"]
known_first_party = ["habits"]
[dependency-groups]
lint = [
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.0,<2.0.0)"
]
dev = [
    "flake8 (>=7.3.0,<8.0.0)"
]
